+++ util/options_helper.cc
@@ -8,12 +8,12 @@
 #include <cstdlib>
 #include <unordered_set>
 #include "rocksdb/cache.h"
+#include "rocksdb/convenience.h"
 #include "rocksdb/filter_policy.h"
 #include "rocksdb/options.h"
 #include "rocksdb/rate_limiter.h"
 #include "rocksdb/slice_transform.h"
 #include "rocksdb/table.h"
-#include "rocksdb/utilities/convenience.h"
 #include "table/block_based_table_factory.h"
 #include "util/logging.h"
 #include "util/options_helper.h"
